Added the alpha version of the map the the static folder.
Trying to standardize the names used for timestamps in the system. Also added some functions to RawData
and EventReport I think "exists()"that is used in the check that prevents duplicates from being saved.

The default time for the google reader task to 1 hr so this may need to be tweaked or so. The google reader
fetches default 60 items each time.